# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Auth API",
        "description": "API для управления аутентификацией пользователей",
        "contact": {
            "email": "gamletel@list.ru"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/comments/{id}": {
            "get": {
                "tags": [
                    "UserComments"
                ],
                "summary": "Получить комментарий по ID",
                "description": "Возвращает комментарий по ID",
                "operationId": "a8a944dc8d89430b9bf9e371b3be0d0f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID комментария",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserComment"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Комментарий не найден"
                    }
                }
            },
            "delete": {
                "tags": [
                    "UserComments"
                ],
                "summary": "Удалить комментарий",
                "description": "Удаляет комментарий по ID",
                "operationId": "0caaadaa7f7471dbe5ff3a695e9a5357",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID комментария",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Комментарий успешно удалён",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Comment is successfully deleted"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Комментарий не найден"
                    }
                }
            }
        },
        "/comments": {
            "post": {
                "tags": [
                    "UserComments"
                ],
                "summary": "Создать новый комментарий",
                "description": "Создает новый комментарий для пользователя",
                "operationId": "2bb76c44f4ecd00f1b534dcb2166b002",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "post_id",
                                    "title",
                                    "text"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "post_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "rating": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Заголовок комментария"
                                    },
                                    "text": {
                                        "type": "string",
                                        "example": "Текст комментария"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Комментарий успешно создан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserComment"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации данных"
                    }
                }
            }
        },
...
        "/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Получить данные пользователя по ID",
                "description": "Возвращает информацию о пользователе по его ID",
                "operationId": "910a0c7cf580e1c605648d82a4e57f46",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID пользователя",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Пользователь не найден"
                    }
                }
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Обновить данные пользователя",
                "description": "Обновляет информацию пользователя по его ID",
                "operationId": "ddfe232db1870b82493bf8bd25596d37",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID пользователя",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "johndoe@example.com"
                                    },
                                    "avatar": {
                                        "type": "string",
                                        "example": "new_avatar.png"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Пользователь обновлён",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/{id}/posts": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Получить посты пользователя",
                "description": "Возвращает все посты пользователя по его ID",
                "operationId": "c581da0de6361e63c0c0c50a36c37971",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID пользователя",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Post"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Post": {
                "description": "Модель поста",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Название поста"
                    },
                    "description": {
                        "type": "string",
                        "example": "Краткое описание"
                    },
                    "text": {
                        "type": "string",
                        "example": "Содержимое поста"
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "icon": {
                        "type": "string",
                        "example": "image.png"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-12-31T12:34:56Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-12-31T12:34:56Z"
                    }
                },
                "type": "object"
            },
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "example": "johndoe@example.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "qweasd"
                    },
                    "avatar": {
                        "type": "string",
                        "example": "image.png"
                    }
                },
                "type": "object"
            },
            "UserComment": {
                "description": "Модель комментария",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "post_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "rating": {
                        "type": "integer",
                        "example": "5"
                    },
                    "title": {
                        "type": "string",
                        "example": "Заголовок"
                    },
                    "text": {
                        "type": "string",
                        "example": "Текст комментария"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-12-31T12:34:56Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-12-31T12:34:56Z"
                    }
                },
                "type": "object"
            }
        }
    },
...
}